backend/requirements.txt
arrow==1.2.2
asgiref==3.5.0
autopep8==1.6.0
binaryornot==0.4.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cookiecutter==1.7.3
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.49.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-grappelli==3.0.3
django-rest-auth==0.9.5
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.1
jinja2-time==0.2.0
MarkupSafe==2.1.1
oauthlib==3.2.0
pep8==1.7.1
poyo==0.5.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-slugify==6.1.1
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.8
